requirements.txt
functions-framework==3.*
wbgapi
pandas
google-cloud-bigquery
pyarrow